Move SDK to packages
