--- .env.example
@@ -1,23 +1,0 @@
-#
-# Runner configuration
-#
-
-# This is a default API key.
-# Runner uses this key to protect the API.
-# Use it in the clients to configure a connection to the runner.
-CONNERY_RUNNER_API_KEY=
-
-# OpenAI API Key is used for the Natural Language Actions feature.
-# It's optional if you don't use the feature.
-OPENAI_API_KEY=
-
-# GitHub Personal Access Token is used to access plugins in private repositories.
-# The user associated with the PAT must have access to the private repositories.
-# It's optional if you don't use private plugins.
-GITHUB_PAT=
-
-# The public URL of the runner. 
-# This URL is used in an OpenAPI specification for the available actions on the runner.
-# It's optional if you don't use the feature.
-# Example: https://runner.example.com (without trailing slash)
-CONNERY_RUNNER_PUBLIC_URL=
