--- packages/_sdk/__package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@connery-io/sdk",
-  "description": "Connery SDK",
-  "version": "0.0.16",
-  "license": "MIT",
-  "dependencies": {},
-  "devDependencies": {
-    "tsconfig": "*"
-  },
-  "scripts": {
-    "build": "tsc",
-    "dev": "tsc -w"
-  }
-}
