+++ packages/sdk/package.json
@@ -7,6 +7,13 @@
   "main": "./dist/index.js",
   "module": "./dist/index.js",
   "types": "./dist/index.d.ts",
+  "scripts": {
+    "dev": "tsc -w",
+    "build": "tsc",
+    "test": "jest --passWithNoTests",
+    "format": "prettier --write .",
+    "lint": "eslint src/**/*.ts"
+  },
   "dependencies": {
     "@nestjs/common": "^9.4.3",
     "@nestjs/config": "^2.3.4",
@@ -36,13 +43,6 @@
     "tsconfig": "*",
     "tsconfig-paths": "4.1.0",
     "typescript": "^4.9.5"
-  },
-  "scripts": {
-    "dev": "tsc -w",
-    "build": "tsc",
-    "test": "jest --passWithNoTests",
-    "format": "prettier --write .",
-    "lint": "eslint src/**/*.ts"
   },
   "jest": {
     "preset": "ts-jest",
