+++ packages/sdk/src/api/controllers/openai.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Get, Inject } from '@nestjs/common';
-import { Public } from './../auth.guard.js';
-import { OpenAiFunctionSchema } from './../../types/llm.js';
+import { Public } from '../auth.guard.js';
+import { OpenAiFunctionSchema } from '../../types/llm.js';
 import { OpenAPIV3 } from 'openapi-types';
 import { OpenAiSpecsService } from '../services/openai-specs.service.js';
