+++ packages/sdk/src/api/controllers/plugins.controller.ts
