+++ packages/sdk/src/api/services/openai-specs.service.ts
