+++ packages/sdk/src/runtime/plugin.ts
@@ -1,5 +1,5 @@
-import { ConfigurationParameterDefinition, MaintainerDefinition, PluginDefinition } from './../types/definition.js';
-import { ActionRuntime, PluginRuntime } from './../types/runtime.js';
+import { ConfigurationParameterDefinition, MaintainerDefinition, PluginDefinition } from '../types/definition.js';
+import { ActionRuntime, PluginRuntime } from '../types/runtime.js';
 import { Action } from './action.js';
 import { validatePluginDefinition } from './plugin-definition-validation-utils.js';
